+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 
 # Change Log
 
-## Unreleased
+## 0.4.0-beta.x
 
 ### ✨ New
 
@@ -22,12 +22,15 @@
     -   Add `streamDeck.settings` namespace for interacting with settings.
     -   Add `streamDeck.system` namespace for system-related operations.
     -   Add `streamDeck.plugin` namespace for bi-direction communication with the plugin and the UI.
+-   Add `isInMultiAction` to the property inspector's action information.
 
 ### 🐞 Fix
 
 -   `Coordinates` type could erroneously have a non-number type for `row`.
 -   Fix support for allowed types within payloads.
 -   Fix localization lookup to index from `Localization`.
+-   Fix race condition when tracking the property inspector.
+-   Fix `streamDeck.setGlobalSettings` to require settings that extend `JsonObject`.
 
 ### ♻️ Update
 
@@ -35,6 +38,9 @@
 -   Localization lookup will now return the key if the resource is not defined.
 -   Update structure of JSON localizations.
 -   Update `State` type to allow for more than two states.
+-   Update routing to prevent exposure of internal messages.
+-   Update build to export Stream Deck API types.
+-   Update `ws` dependency.
 
 ### ⬆️ Upgrading